+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Releases
 
-- **Added:** Add `project.urls` section to `pyproject.toml` 
+- **Added:** Support for checking variables in docstring sections
+- **Added:** project.urls` section to `pyproject.toml` 
 - **Added:** Support for Module and Class docstring checks.
 - **Fixed:** `DOC006` false positive.